refactor: move issue parsers to their own files
